+++ DTCD-HeaderPanel/src/Plugin.js
@@ -11,12 +11,14 @@ import {
   AppGUISystemAdapter,
   NotificationSystemAdapter,
   LogSystemAdapter,
+  StorageSystemAdapter,
 } from './../../DTCD-SDK/index';
 
 export class Plugin extends AppPanelPlugin {
   #vue;
   #workspaceSystem;
   #interactionSystem;
+  #storageSystem;
   #logSystem;
   #idAuthorizedUser = null;
   static userEndpoint = '/dtcd_utils/v1/user?photo_quality=low';
@@ -40,6 +42,7 @@ export class Plugin extends AppPanelPlugin {
     const eventSystem = new EventSystemAdapter('0.4.0', guid);
     this.#interactionSystem = new InteractionSystemAdapter('0.4.0');
     this.#workspaceSystem = new WorkspaceSystemAdapter('0.4.0');
+    this.#storageSystem = new StorageSystemAdapter('0.9.0');
     const styleSystem = new StyleSystemAdapter('0.4.0');
     const router = new RouteSystemAdapter('0.1.0');
     const appGUI = new AppGUISystemAdapter('0.1.0');
@@ -65,6 +68,7 @@ export class Plugin extends AppPanelPlugin {
       interactionSystem: this.#interactionSystem,
       eventSystem,
       workspaceSystem: this.#workspaceSystem,
+      storageSystem: this.#storageSystem,
       styleSystem,
       router,
       appGUI,