+++ DTCD-HeaderPanel/src/components/NotificationBell.vue
@@ -4,24 +4,24 @@
     @click="keepOpen = 2"
   >
     <button
-      class="button-icon"
+      class="IconButton"
       @click.prevent="clickOnIcon()"
     >
       <span
-        class="FontIcon name_notification"
+        class="FontIcon name_notification size_md"
         :class="iconClasses"
       ></span>
     </button>
 
     <div
       v-if="showList"
-      class="notification-scrolling"
+      class="NotificationContainer"
     >
-      <div class="notification-list">
-        <div class="btn-header">
+      <div class="NotificationList">
+        <div class="HeaderButton">
           <base-button
             v-if="notifications.length >= 5"
-            class="clear-btn"
+            class="ButtonClear"
             @click.prevent="$root.notificationSystem.clearList()"
             theme="theme_blueSec"
           >
@@ -30,7 +30,7 @@
         </div>
         <div
           v-if="notificationList.length === 0"
-          class="empty-text"
+          class="EmptyNotificationList"
         >
           Уведомлений нет
         </div>
@@ -41,7 +41,7 @@
           <div
             v-for="{ title, body, className, id, hasAction } of notificationList"
             :key="id"
-            class="notification-item"
+            class="NotificationItem"
             :class="className"
           >
             <span 
@@ -51,12 +51,12 @@
             </span>
             <span 
               @click.prevent="$root.notificationSystem.remove(id)" 
-              class="FontIcon name_closeBig size_md close-btn"
+              class="FontIcon name_closeBig size_md ButtonClose"
             >
             </span>
             <div
               v-if="title" 
-              class="title"
+              class="NotificationTitle"
               :class="{
               'has-action': hasAction,
               }"
@@ -71,7 +71,7 @@
             </div>
             <div 
               v-if="body" 
-              class="body-text"
+              class="NotificationBody"
             >
               <vue-show-more-text
                 :text="body"
@@ -87,7 +87,7 @@
       </div>
     </div>
 
-    <div :class="`notification-list floating-list ${settings.notificationPosition}`">
+    <div :class="`NotificationList type_floating ${settings.notificationPosition}`">
       <transition-group 
         name="list" 
         tag="div"
@@ -95,9 +95,10 @@
         <div
           v-for="item of notificationFloatList"
           :key="item.id"
-          class="notification-item"
+          class="NotificationItem"
           :class="item.className"
           @mouseenter="onMouseEnterFloatItem(item)"
+          @mouseleave="onMouseLeaveFloatItem(item)"
         >
           <span 
             class="FontIcon FrontIcon"
@@ -106,12 +107,12 @@
           </span>
           <span 
             @click.prevent="$root.notificationSystem.remove(item.id)"
-            class="FontIcon name_closeBig size_md close-btn"
+            class="FontIcon name_closeBig size_md ButtonClose"
           >
           </span>
           <div
             v-if="item.title"
-            class="title"
+            class="NotificationTitle"
             :class="{
             'has-action': item.hasAction,
             }"
@@ -126,7 +127,7 @@
           </div>
           <div 
             v-if="item.body" 
-            class="body-text"
+            class="NotificationBody"
           >
             <vue-show-more-text
               :text="item.body"
@@ -201,13 +202,21 @@ export default {
       const {
         notifications,
       } = this;
-      return {
-        indication: !!notifications.length,
-        'has-success': notifications.findIndex(({options}) => options.type === 'success') !== -1,
-        'has-warning': notifications.findIndex(({options}) => options.type === 'warning') !== -1,
-        'has-error': notifications.findIndex(({options}) => options.type === 'error') !== -1,
-      };
-    },
+
+      let className = !!notifications.length ? 'indication' : '';
+
+      if (notifications.findIndex(({options}) => options.type === 'error') !== -1) {
+        return `${className} type_error`
+      }
+      if (notifications.findIndex(({options}) => options.type === 'warning') !== -1) {
+        return `${className} type_warning`
+      }
+      if (notifications.findIndex(({options}) => options.type === 'success') !== -1) {
+        return `${className} type_success`
+      }
+
+      return `${className}`
+    }, 
   },
   mounted() {
     this.notifications = this.$root.notificationSystem.getList();
@@ -269,11 +278,15 @@ export default {
       }
 
       if (options.floatMode) {
-        options.floatTimeOut = setTimeout(() => {
-          options.floatMode = false;
-          options.floatTimeOut = undefined;
-        }, (options.floatTime || 5) * 1000)
+        this.setTimer(options);
       }
+    },
+
+    setTimer(options)  {
+      options.floatTimeOut = setTimeout(() => {
+        options.floatMode = false;
+        options.floatTimeOut = undefined;
+      }, (options.floatTime || 5) * 1000)
     },
 
     removeNotification(id) {
@@ -295,11 +308,15 @@ export default {
     },
 
     onMouseEnterFloatItem(item) {
-      if (item.options.floatTimeOut) {
+      if (item.options.floatTimeOut) {  
         clearTimeout(item.options.floatTimeOut);
         item.options.floatTimeOut = undefined;
       }
     },
+
+    onMouseLeaveFloatItem(item) {
+      this.setTimer(item.options);
+    }
   }
 }
 </script>
@@ -324,13 +341,13 @@ export default {
     box-sizing: border-box;
   }
 
-  .button-icon {
+  .IconButton {
     background: transparent;
     border: none;
     cursor: pointer;
   }
 
-  .notification-scrolling {
+  .NotificationContainer {
     position: absolute;
     right: 0;
     top: 30px;
@@ -352,28 +369,29 @@ export default {
     }
   }
 
-  .btn-header {
+  .HeaderButton {
     position: sticky;
     top: 0;
     background-color: var(--background_secondary);
     text-align: center;
     z-index: 1;
-    padding-bottom: 10px;
+    padding-bottom: 5px;
     margin-bottom: -5px;
   }
 
-  .clear-btn {
+  .ButtonClear {
     margin-top: 10px;
+    margin-bottom: 5px;
     width: 80%;
   }
 
-  .empty-text {
+  .EmptyNotificationList {
     padding: 32px 16px;
     text-align: center;
     color: var(--text_main);
   }
 
-  .notification-list {
+  .NotificationList {
     background-color: var(--background_secondary);
     border: 1px solid var(--border);
     border-radius: 8px;
@@ -384,7 +402,7 @@ export default {
     width: 360px;
     margin-right: -8px;
 
-    &.floating-list {
+    &.type_floating {
       position: fixed;
       z-index: 30;
       padding: 0;
@@ -394,7 +412,7 @@ export default {
       box-shadow: none;
       max-height: 100%;
       right: 0;
-      bottom: 0;
+      bottom: 20px;
       &.top {
         bottom: auto;
         top: 20px;
@@ -417,84 +435,85 @@ export default {
       border: 4px solid transparent;
       background-clip: content-box;
     }
+  }
 
-    .notification-item {
-      margin: 6px;
-      padding: 10px;
-      background: var(--border_12);
-      border-radius: 8px;
-      border: 1px solid;
-      color: var(--accent);
-      position: relative;
-      min-height: 20px;
+  .NotificationItem {
+    margin: 6px;
+    padding: 10px;
+    background: var(--border_12);
+    border-radius: 8px;
+    border: 1px solid;
+    color: var(--accent);
+    position: relative;
+    min-height: 20px;
 
-      .FrontIcon {
-        position: absolute;
-        left: 0;
-        top: 0;
-        margin: 10px;
-        display: flex;
-        font-size: 18px;
-        align-items: center;
+    &.success {
+      color: var(--success);
+
+      .FontIcon {
+        &.name_circleCheckOutline {
+          color: var(--success);
+        }   
       }
+    }
 
-      .close-btn {
-        position: absolute;
-        right: 0;
-        top: 0;
-        margin: 10px;
-        text-decoration: none;
-        cursor: pointer;
+    &.warning {
+      color: var(--warning);
 
-        &:hover {
-          color: var(--text_main);
-        }
+      .FontIcon {
+        &.name_warningOutline {
+          color: var(--warning);
+        }   
       }
+    }
 
-      .title {
-        font-weight: 700;
-        font-size: 16px;
-        position: relative;
-        margin: 0 18px 0 21px;
-      }
+    &.error {
+      color: var(--danger);
 
-      .body-text {
-        margin-top: 8px;
-        font-size: 14px;
-        color: var(--text_main);
-      }
-
-      &.success {
-        color: var(--success);
-
-        .FontIcon {
-          &.name_circleCheckOutline {
-            color: var(--success);
-          }   
-        }
-      }
-
-      &.warning {
-        color: var(--warning);
-
-        .FontIcon {
-          &.name_warningOutline {
-            color: var(--warning);
-          }   
-        }
-      }
-
-      &.error {
-        color: var(--danger);
-
-        .FontIcon {
-          &.name_offOutlineClose {
-            color: var(--danger);
-          }   
-        }
+      .FontIcon {
+        &.name_offOutlineClose {
+          color: var(--danger);
+        }   
       }
     }
   }
+    
+  .FrontIcon {
+    position: absolute;
+    left: 0;
+    top: 0;
+    margin: 10px;
+    display: flex;
+    font-size: 18px;
+    align-items: center;
+  }
+
+  .ButtonClose {
+    position: absolute;
+    right: 0;
+    top: 0;
+    margin: 10px;
+    text-decoration: none;
+    cursor: pointer;
+
+    &:hover {
+      color: var(--text_main);
+    }
+  }
+
+  .NotificationTitle {
+    font-weight: 700;
+    font-size: 16px;
+    position: relative;
+    margin: 0 18px 0 21px;
+  }
+
+  .NotificationBody {
+    margin-top: 8px;
+    font-size: 14px;
+    color: var(--text_main);
+  }
+  
   .FontIcon {
     color: var(--accent);
 
@@ -504,11 +523,7 @@ export default {
 
     &.name_notification {
       position: relative;
-
-      &::before {
-        font-size: 18px;
-        color: var(--general_white);
-      }
+      color: var(--general_white);
 
       &.indication {
         &::after {
@@ -523,20 +538,19 @@ export default {
           top: 0;
         }
 
-        &.has-success {
+        &.type_success {
           &::after {
             background-color: var(--success);
           }
         }
 
-        &.has-warning {
+        &.type_warning {
           &::after {
             background-color: var(--warning);
           }
-          color: var(--warning);
         }
 
-        &.has-error {
+        &.type_error {
           &::after {
             background-color: var(--danger);
           }