+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - saving setting of notification position to localStorage
 
+### Changed
+
+- margin and padding for deletes all notifications button
+
 ## [0.6.1]
 
 ### Added
@@ -21,7 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - design of notifications
 - text icons to font icons
-- margin and padding for deletes all notifications button
 
 ## [0.5.0]