+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - plugins version sorting
 - interface for copying dashboard link
 
+### Changed
+
+- positioning of bottom notifications
+
 ### Fixed
 
 - notification behavior after mouseover