+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - interface for copying dashboard link
 - download logo
 - only priority notification class is reflected
+- displaying of workspace title when it's opened
 
 ### Changed